+++ plugins/mesh/emit/object/emit.h
@@ -439,7 +439,7 @@ public:
   // Redirect some iEmitState functions to the csParticleSystem implemenation
   virtual void SetParticleCount (int num)
   { csParticleSystem::SetParticleCount (num); }
-  virtual int GetParticleCount () const
+  virtual size_t GetParticleCount () const
   { return csParticleSystem::GetParticleCount (); }
 };
